pyproject.toml
[tool.poetry]
name = "currency-parcer"
version = "0.1.0"
description = ""
authors = ["ksredkin <shprotyk17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.5"
python-dotenv = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
